agents/agent-runtime/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pyyaml>=6.0.1
python-multipart>=0.0.6
# Document extraction
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.2
